+++ src/strategies/component.ts
@@ -1,54 +1,17 @@
 // import renderer from "./renderer.ts";
-import { component, sibling } from "./types.ts";
+import { component } from "./types.ts";
 
 function Component(this: component, label: string, path: string) {
   this.label = label;
   this.path = path;
   this.child = null;
   this.sibling = null;
-}
-
-// interface storage {
-//   App: component;
-//   Green: component;
-//   Purple: component;
-//   Orange: component;
-// }
-
-/* storage <-- populated via walk with "label" as property names with a new
-component w/ (label, path) as values */
-// prettier-ignore
-const Storage: any = {
-  "App": new (Component as any)("App", "./App.vue"),
-  "Green": new (Component as any)("Green", "./components/Green.vue"),
-  "Purple": new (Component as any)("Purple", "./components/Purple.vue"),
-  "Orange": new (Component as any)("Orange", "./components/Orange.vue"),
 };
 
-function SiblingList(this: sibling) {
-  this.head = null;
-  this.tail = null;
-}
+export default Component;
 
-/**
- * if list has no current head, the initial descendent (child) passed into it
- * will become the head and tail. Otherwise, the tail's sibling (ie next) will become the
- * descendent and we reassign tail to the descendent
- */
-SiblingList.prototype.add = function (descendent: component) {
-  if (!this.head) {
-    this.head = descendent;
-    this.tail = descendent;
-  } else {
-    this.tail.sibling = descendent;
-    this.tail = descendent;
-  }
-};
-const komponents: any = ["Green", "Purple", "Orange"];
-// "Green"
-
+/*
 function listBuilder(parent: component) {
-  // this = App
   parent.child = new (SiblingList as any)();
   while (komponents.length) {
     const str = komponents.pop(); // <-- currently "Orange"
@@ -57,16 +20,9 @@ function listBuilder(parent: component) {
   }
   console.log("sibling list", parent.child);
 }
+*/
 
-const { App } = Storage;
 
-listBuilder(App);
-
-// const current = komponents.pop();
-// parent.child.sibling = current;
-// const current = storage[komponents.pop()]
-
-// listBuilder(storage.App)
 /**
  * parser will contain the components of the tree:: cache will be the tree
  *
@@ -77,29 +33,3 @@ listBuilder(App);
  
  * queue = [ Green, Purple, Orange ]
 */
-/**
- * function LinkedList() {
-  this.head = null;
-  this.tail = null;
-  this.length = 0;
-}
-
-function Node(val) {
-  this.value = val;
-  this.next = null;
-}
-
-LinkedList.prototype.push = function (value) {
-  const node1 = new Node(value);
-  // if there is no head or tail in the Linkedlist, then the node just created is the new head and tail
-  if (!this.head) {
-    this.head = node1;
-    this.tail = node1;
-  } else {
-    // if there is a head, the new node that is created is the new tail and we must reassign the value of "next" for old node to point to the new tail
-    this.tail.next = node1;
-    this.tail = node1;
-  }
-  this.length += 1;
-};
- */
