+++ src/strategies/storage.ts
@@ -1,0 +1,10 @@
+import { component } from "./types.ts";
+
+interface storage {
+  [key: string]: component;
+ 
+}
+interface StorageConstructor {};
+
+const Storage = function (this: storage): storage {
+};
